server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow
PyMySQL
Flask-JWT-Extended
python-dotenv
bcrypt
